3d TMs
UPF files for the TMs are included. Each solved for a +2 reference
state, where the state is 3s2 3p 3dn. This achieves reasonable
transferability for states near this oxidation state. If the system is
far from this oxidation state, the projector may not be as transferable.
Check the TM.ccECP.rpt file for the transferability information and
planewave cutoffs