Use "whereis" instead of "which" for dependency checks
As a result of recent updates, the "which" command is no longer returning the expected path string when executed via "exec_shell". This resulted in falsely thrown errors when deploy.php was executed. The "whereis" command returns predictable output and therefore it can be used instead of "where" to correct the existing deployment script with minimal changes.